requirements-dev.txt
black
flake8-bugbear
flask-debugtoolbar
flask-shell-ipython
ipdb
ipython
isort
pytest-flask
pytest-cov
 